+++ handlers/kinetic_request_ce_attachment_create_v1/CHANGELOG.md
@@ -1,3 +1,6 @@
+Kinetic Request CE Attachment Create V1.3 (2021-05-27)
+ * Added the ability to accept Base 64 input.
+
 Kinetic Request CE Attachment Create V1.2 (2018-05-25)
 * API Server Info Value changed to allow ${space} in the url for subdomain support
 (ie. https://${space}.localhost:8080/kinetic)